test/vhost: do not overwrite fio config file
Fio output was saved to the same file as original
fio config. For a single test run this was not an issue,
but in case we run more iterations (like in performance
tests) we cannot overwrite the original config file.

Use fio config file name (without file extension) and add
".log" to the name.

Signed-off-by: Karol Latecki <karol.latecki@intel.com>
Change-Id: If6f9d582ff5817c85b7cbf9192530b87a2079274
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/1898
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>