test/lvol: [jq] replace -S with sort()
It seems like -S will sort only keys within given objects, but if
actual objects are in different order, the sorting results may still
differ.

To avoid this, replace -S with jq's sort() which takes entire array,
and all its objects, as an input to be sorted. This should result in
the same output regardless of the initial position of given objects.

Change-Id: I4eca9474f1cea73040ae4a13692a1074fa1c4fc3
Signed-off-by: Michal Berger <michalx.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/1973
Reviewed-by: Vitaliy Mysak <vitaliy.mysak@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>